requirements.txt
numpy
requests
tqdm
scikit-learn
pandas
langchain==0.3.19
regex
trl==0.8.2
peft
accelerate
safetensors>=0.4.1
datasets==2.18.0
torch>=2.2.0
torchaudio
transformers
bitsandbytes
wandb
langchain-community==0.3.17
huggingface_hub
openai
python-dotenv
langgraph==0.2.74
pytest
spacy==3.7.5
spacy-transformers==1.3.8
es-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/es_core_news_sm-3.7.0/es_core_news_sm-3.7.0-py3-none-any.whl
fuzzywuzzy
python-Levenshtein
fastapi
uvicorn
pydantic
beautifulsoup4
langdetect
pandasql
sqlalchemy
aiohttp
tree
openpyxl
langchain-openai==0.3.6
sentence-transformers>=2.6.0
langchain-huggingface==0.1.2
pytest-asyncio
faiss-cpu
psycopg2-binary
prometheus_client
pytest-asyncio
asyncpg
PyPDF2
argostranslate
openai>=1.0.0
anthropic>=0.7.0